Update README.md
This project automates the end-to-end provisioning and configuration of a web server on AWS. Terraform is used to define and deploy infrastructure components like VPCs, subnets, security groups, and EC2 instances in a repeatable, version-controlled way. Once the infrastructure is live, Ansible is used for post-deployment configuration, including package installation, service setup, and firewall rules. The goal of the project is to demonstrate infrastructure as code (IaC) principles, improve DevOps workflow efficiency, and practice secure and scalable cloud architecture.